Usb package model can now setted to null
If the user no longer wants to track a mission, the usb manager service will no longer be able to analyze the received data. That way, if a usb device is plugged in and there is no mission to track, the usb will not provide packages for the rest of the app.