Using a document's creation date as its own name
When a new document is loaded into Firebase, it uses the randomly generated document name to sort the entire collection alphabetically. So if we wanted to download the documents and show them based on the creation date, we would have to sort the documents every time they were downloaded. Using the document created date as the name, Firebase will maintain the correct order of documents based on their creation.